# component/file/api/v1/file.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "component/file/api/v1/file.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "FileService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/files": {
      "get": {
        "operationId": "FileService_ListFiles",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListFilesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "sysId",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "catId",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "FileService"
        ]
      },
      "post": {
        "operationId": "FileService_CreateFile",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1File"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateFileRequest"
            }
          }
        ],
        "tags": [
          "FileService"
        ]
      }
    },
    "/api/v1/files/{id}": {
      "get": {
        "operationId": "FileService_GetFile",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1File"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "FileService"
        ]
      },
      "delete": {
        "operationId": "FileService_DeleteFile",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "FileService"
        ]
      },
      "put": {
        "operationId": "FileService_UpdateFile",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1RowsAffected"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "file": {
                  "$ref": "#/definitions/v1File"
                },
                "updateMask": {
                  "type": "string"
                }
              },
              "required": [
                "file"
              ]
            }
          }
        ],
        "tags": [
          "FileService"
        ]
      }
    }
  },
  "definitions": {
    "FileFileStatus": {
      "type": "string",
      "enum": [
        "StatusDefault",
        "StatusActive",
        "StatusDisable"
      ],
      "default": "StatusDefault"
    },
    "FileFileType": {
      "type": "string",
      "enum": [
        "FileTypeDefault",
        "FileTypeImage",
        "FileTypeVideo",
        "FileTypeDoc",
        "FileTypeOther"
      ],
      "default": "FileTypeDefault"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1CreateFileRequest": {
      "type": "object",
      "properties": {
        "parent": {
          "type": "string"
        },
        "file": {
          "$ref": "#/definitions/v1File"
        },
        "fileId": {
          "type": "string"
        }
      },
      "required": [
        "file"
      ]
    },
    "v1File": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "sysId": {
          "type": "string",
          "format": "int64"
        },
        "catId": {
          "type": "string",
          "format": "int64"
        },
        "itemId": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string",
          "format": "int64"
        },
        "fileType": {
          "$ref": "#/definitions/FileFileType"
        },
        "name": {
          "type": "string"
        },
        "ext": {
          "type": "string"
        },
        "path": {
          "type": "string"
        },
        "hash": {
          "type": "string"
        },
        "status": {
          "$ref": "#/definitions/FileFileStatus"
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "Output only. Creation time of the file.",
          "readOnly": true
        },
        "updateTime": {
          "type": "string",
          "format": "date-time"
        },
        "deleteTime": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "v1ListFilesResponse": {
      "type": "object",
      "properties": {
        "total": {
          "type": "string",
          "format": "int64"
        },
        "totalPages": {
          "type": "string",
          "format": "int64"
        },
        "pageSize": {
          "type": "string",
          "format": "int64"
        },
        "currentPage": {
          "type": "string",
          "format": "int64"
        },
        "files": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1File"
          }
        }
      }
    },
    "v1RowsAffected": {
      "type": "object",
      "properties": {
        "rowsAffected": {
          "type": "string",
          "format": "int64"
        }
      }
    }
  }
}